+++ src/fichier_json/version.json
@@ -1,8 +1,7 @@
 {"data":[
-{"nom":"clio","id":"ceAD233W6ZuOPn6vX33L","url":"qsd"},
-{"id":"lsq8mz9Un09gSs51Sbel","url":"qsdfqsf","nom":"congo"},
-{"id":"oBQKWEqtQ0PbTnolDeLq","url":"qz","nom":"symbol"},
-{"id":"2zUhjue9F8pDEaDqhp5B","url":"qsdg","nom":"fluence"},
-{"id":"uevaaezw7coOKxHPvHqF","url":"world","nom":"hello"}
+{"id":"ceAD233W6ZuOPn6vX33L","nom":"1.0","prix":1000000,"options":[ops1,ops3,ops5]},
+{"id":"lsq8mz9Un09gSs51Sbel","nom":"1.2","prix":1500000,"options":[ops1,ops2,ops4]},
+{"id":"oBQKWEqtQ0PbTnolDeLq","nom":"2.2","prix":1200000,"options":[ops1,ops2,ops3]},
+{"id":"2zUhjue9F8pDEaDqhp5B","nom":"2.2","prix":2000000,"options":[ops1,ops2,ops3]}
 ]
 }
